+++ srcs/redirect/ft_is_original_fd.c
@@ -3,16 +3,16 @@
 /*                                                        :::      ::::::::   */
 /*   ft_is_original_fd.c                                :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: bmarilli <marvin@42.fr>                    +#+  +:+       +#+        */
+/*   By: exenia <exenia@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/04/02 03:23:17 by bmarilli          #+#    #+#             */
-/*   Updated: 2021/04/07 23:25:31 by bmarilli         ###   ########.fr       */
+/*   Updated: 2021/04/12 20:13:46 by exenia           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include <minishell.h>
 
-int	ft_is_original_fd(int *fds)
+int	ft_is_original_fd(long long int *fds)
 {
 	if (fds[0] == 0 && fds[1] == 1 && fds[2] == 2)
 		return (1);
